# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsArticle {
  id: number;
  headline: string;
  category: string;
  bullets: string[];
  imageUrl?: string | null;
  publishedAt: string;
  processedAt: string;
}

export type ListNewsParams = {
  category?: string;
  limit?: number;
};