# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: news
    description: News article operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List processed news articles
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get one article with bullet points
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
  /news/categories:
    get:
      operationId: listCategories
      tags: [news]
      summary: List available categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    NewsArticle:
      type: object
      properties:
        id:
          type: integer
        headline:
          type: string
        category:
          type: string
        bullets:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
      required:
        - id
        - headline
        - category
        - bullets
        - publishedAt
        - processedAt