# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: number;
  headline: string;
  category: string;
  bullets: string[];
  imageUrl?: string | null;
  publishedAt: Date;
  processedAt: Date;
}